[FIX] documents_spreadsheet: support Reference fields in pivot side panel
Steps to reproduce the issue on a fresh runbot database:
- open the Subscription dashboard in edit mode
- Right click on cell "Data!B7"
=> boom

The pivot `#11` measure is `order_reference` which is a Reference field.
This type of field had no aggregator specified which crashes the side panel.

closes odoo/enterprise#60738

Note: This commit is only a UI test. See the fix in the community branch.
Task: 3869693
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>