[FIX] l10n_mx_edi,*: Manage approval/rejection of cancellation request
Case 1:
The user can cancel the document directly from the SAT.
In that case, in odoo, the user needs to "Request Cancel" first
even if the SAT state becomes 'cancelled'.
To improve that, this case is now managed by adding a new cancel
document automatically on the invoice.

Case 2:
The user can request a cancellation from the SAT.
Currently, the invoice is marked as "cancel" even if the SAT rejects the cancellation.
In order to manage that, let's introduce a new document: 'invoice_cancel_requested'.

closes odoo/enterprise#60905

Signed-off-by: Claire Bretton (clbr) <clbr@odoo.com>