[FIX] sale_timesheet_enterprise: fix wrong indicator in the leaderboard
If the current user isn't in the leaderboard, a message gets displayed
on the indicator: "Record timesheets to earn your rank!" or "Record
timesheets to determine your billing rate!", depending on if the
leaderboard feature is on or not. The condition for this was wrong, and
as such, the second message would get displayed whether the feature was
on or not. This PR fixes the condition.

closes odoo/enterprise#60755

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>