[FIX] documents: make chatter flex in Documents app
The chatter in documents was not flex, which made it take a lot
of space without any wrapping. As a result, usually chatter took
all the screen and content was massively overflowing, resulting
in poor UX.

This was caused by a specific stylerule in documents with chatter
that made sense in a earlier version of chatter CSS, but this
is no longer needed.

Also we actually want to reuse most style of chatter in form view.
This commit adds `o-mail-ChatterContainer` classname on same
HTML node as `o-mail-Form-Chatter` and adapts style, so that
Document can set this classname to reuse style.

opw-3681435

closes odoo/enterprise#61330

X-original-commit: a2d2c78d9a5091784b460bbf45e598100f0f086a
Related: odoo/odoo#163026
Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>