[IMP] l10n_au_hr_payroll: Fixes Computation issues for unused leaves
The amount withheld from the unused leaves payment at the time of
termination depends on the normal gross earnings of an employee
during their schedule pay period. This commit implements
withholding variations for unused leaves. Also adds
the option to assign the gross unused leaves manually.

Issue: The salary rules use the gross salary
computed on a regular payslip. This raises the issues
when the employee is terminated in the middle of a
pay period and the gross salary for the payslip
changes.

task# 3821959

closes odoo/enterprise#61736

Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>