[FIX] project_sale_subscription: create recurring task from portal
Steps to reproduce:
-------------------
- create a product with:
    - recurring
    - prepaid
    - create task on order
- in the settings, enable "Recurring Tasks"
- add "Use Recurring Tasks" to the current user
- create a sale order with a recurring plan and the recurring product
- confirm the sale order
   --> the created task is recurring
- duplicate the sale order
- validate it with a portal user

Issue:
------
The task created is not recurring.
The two tasks should be recurrent because the sale order is the same.
The user who confirms the sale order must not affect the status of the task.

Cause:
------
We check that the user belongs to the `group_project_recurring_tasks` group,
which will never be the case for a portal user.

Solution:
---------
Authorise the creation of the recurring task if the setting is activated.

opw-3823250

closes odoo/enterprise#60807

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>