[IMP] account_reports: default filter
In the balance sheet and Profit & Loss reports, we allow users to audit the
General Ledger of a specific account if we want to look at the details of that
account for the period of the report. When clicking on the link of the audit,
it works well for small db, but if the account is low in the list, you don't
even see it. This commit adds a default filter on the report.

task-3839845

closes odoo/enterprise#63931

X-original-commit: 7135850d407a2bc3fcbc7bcee702de44ea2bf912
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>
Signed-off-by: Habib Ayob (ayh) <ayh@odoo.com>