[IMP] account_fleet : add link to bill in fleet
With this commit, the bill's amount is correctly computed.
If the service log is generated by a bill, the cost is readonly to be
consistent.
If the service log is generated by a bill, it can't be deleted.

The bill smartbutton's color is related to the bill's state.

task-3783017

closes odoo/enterprise#60818

Related: odoo/odoo#157245
Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>