[IMP] appointment: reorder gantt form, update duration of appointment
Improve the form view of calendar event opened on the gantt
view of appointments.

1) As many user would use the gantt view (general one with all
resources, all users) when working daily with appointments,
add a small field widget that allows on the gantt form view to
update the duration when one changes the appointment type, as
long as the duration on opening the form is the same as the one
of the appointment type (could be different as one can select
different intervals on the gantt view)

2) Reorder fields to have fields related to the duration come
before the duration field itself.

3) Select Date separately for start and stop instead of using
the daterange widget.

Task-3861242

closes odoo/enterprise#60838

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>