[FIX] l10n_cl_edi_pos: handle missing barcode during order validation
Prior to this commit, an error would occur when validating an order if
the account move associated with a journal (where
'l10n_latam_use_documents' is set to false) did not have
the 'l10n_cl_sii_barcode'.

This commit prevents this error by adding appropriate handling for
cases where the barcode is missing.

opw-3870174

closes odoo/enterprise#60842

Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>