[ADD] account_accountant_check_printing: Add check number to bank rec
desired behavior: If Checks are activated in the Accounting settings, add a new column in the Bank Reconciliation widget
for payment_id.check_number, in optional=hide

Reason: the Check Number is a very useful information to display in the Bank Reconciliation.

solution: create new module dependent on accounting and account_check_printing modules and add the desired column and method there.

Task-3869134

closes odoo/enterprise#64618

X-original-commit: be810e13a2344e8f80eb2e8d16610d4ebceb1e09
Signed-off-by: Quentin De Paoli <qdp@odoo.com>
Signed-off-by: Muhamed Abdalla (muab) <muab@odoo.com>