[FIX] iot: Worldline duplicate Odoo instance fixed
Currently, when downloading/loading Worldline lib at the start of Odoo,
after doing so we are restarting Odoo at the end of
"load_worldline_library.sh" script.

This is unnecessary as Odoo is capable of detecting the terminals and
communicating with them straight away when it has the library and the
linker set up.

More annoyingly, this leads to a very tricky bug:
Odoo service is being restarted in a subshell.
Since Odoo is already running this 2nd service will be blocked, but as
soon as the primary Odoo service stops, the 2nd one takes the lead.

So when restarting Odoo we can get error like "Adress already in use",
duplicate code executed and even more dangerousely, duplicate C++/C code
running with Worldline and Six, which can lead to race conditions and
crashes.

task-3873045

closes odoo/enterprise#60990

X-original-commit: 0874ef0878bae8cf380510e7a8b8295e3ccc0768
Signed-off-by: Quentin Lejeune (qle) <qle@odoo.com>