[FIX] mrp_workorder: set group_mrp_routings to true in tours test
Issue:

Tours from `tour_shopfloor.js` class fail in "No demo" databases.

Explanation:

This is due to the setting `group_mrp_routings` being disabled.
When it is enabled, a pop-up appears when entering Shop Floor and the tours take it into account since the setting is enabled with Demo Data.
When it is disabled, the pop-up does not appear, and the tests fail at the first step because of it.
https://github.com/odoo/enterprise/blob/4b34efc77b57562e4c92956c63c8c20e6de53bf0/mrp_workorder/static/src/mrp_display/mrp_display.js#L107
https://github.com/odoo/enterprise/blob/4b34efc77b57562e4c92956c63c8c20e6de53bf0/mrp_workorder/static/src/mrp_display/mrp_display.js#L114-L119

Fix:

Due to both tests failing for the same reason, the setting will automatically be enabled in `setUpClass`.
`test_shop_floor` also needs an employee named 'Marc Demo', adding the creation of the employee for "No demo" databases.
https://github.com/odoo/enterprise/blob/74e0f4fec69d6512e213312bf42f3d743f7a27b3/mrp_workorder/static/tests/tours/tour_shopfloor.js#L28-L31

error-60565
error-58036

closes odoo/enterprise#60868

Signed-off-by: William Henrotin (whe) <whe@odoo.com>