[FIX] sale_subscription: allow to set to draft
Before this commit, when a confirmed order was canceled, a draft quote
subscription_state was set. It would trigger the constraint
`sale_subscription_state_coherence` or _constraint_canceled_subscription
depnding the version.

THis commits ensure to clean the subscription_state of canceled
subscription and it set it back to the correct value if the canceled
order is set back to quotation.

closes odoo/enterprise#60908

Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>