[IMP] {website_}appointment{_sale}: check upcoming appointment
Add an "extra" step to tell the user that he already had an
upcoming appointment. The idea is to be sure that the user
knows that he already booked an appointment before booking
another one.

The upcoming appointment info is based on the info stored in the
localStorage when an appointment is taken and landing on the
validation page and the sale confirmation if website_appointment_sale
is installed.

In case it's not the public user, we don't use the list of access_token
from the local storage and just check the appointments based on the
appointment_booker_id.

There are 2 buttons, the first one allows to reach the upcoming
appointment details and the other to display the calendar to book
another appointment.
Clicking on the button "Add another" add a value in the local storage
to ignore and don't show the upcoming block info for a day.

We also add the upcoming events info of the appointment booker in
the mail template.

task-3773579

closes odoo/enterprise#61740

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>