[ADD] l10n_uk_{hmrc,cis}: Added Construction Industry Scheme for UK
When a company is employing partner for construction work a special report mechanism is needed.
A Subcontractor is a partner that will do construction work and invoice the Contractor. Construction Industry Scheme add deductions for labour cost on the invoice.
The deducted amount must be paid and reported by the contractor through a monthly return.

For more informations about CIS, see: https://www.gov.uk/what-is-the-construction-industry-scheme

This commit adds:
	- A new API that enables us to use the HMRC transaction engine
	- 6 new taxes for deductions 0, 20, 30% in sales and purchase
	- A CIS report inherited from the tax report
	- A way to directly send the report to hmrc using the transaction engine
	- New fields on res.company and res.partner for the HMRC API

task:3851467

closes odoo/enterprise#66766

Signed-off-by: Olivier Colson (oco) <oco@odoo.com>