[FIX] l10n_de_reports: rounding error adjustment in datev export
Configure 'Cash Discount Tax Reduction' on 'On early payment'
Have a payment term configured with 2% discount for early payment
Create an invoice with the said payment terms and the following lines:
1) qty 1, price unit 468, discount 28, tax 19%
2) qty 1, price unit 480, discount 28, tax 19%
3) qty 1, price unit 85, discount 28, tax 19%
4) qty 1, price unit 6, discount 0, tax 19%
5) qty 2, price unit 6, discount 0, tax 19%
Register the payment with the early payment discount
Check the payment
Early payment lines amount are 15.23 + 2.90 (tax discount) = 18.13
Go to General Ledger
Export "DATEV (ZIP)" and check accounting_entries.csv
Early payment line amount will be 18.12

This occurs because when exporting the system recompute the tax amount
due for each line with tax, so we give just a line with the gross
amount.
However, when the payment is created, the system check for rounding
errors and adjust the computed tax amount. This is not done when
exporting data and we have a 0.01 difference with the payment entry

opw-3801374

closes odoo/enterprise#62006

X-original-commit: acc2cd23b3b19fa4f5eb696537780cd5c5cc2dee
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>
Signed-off-by: Andrea Grazioso (agr) <agr@odoo.com>