[FIX] hr_appraisal_survey: correct the rule
The survey.survey_user_input_rule_survey_user_read rule is
override in both hr_appraisal_survey and in hr_recruitment_survey.

The problem arises when both modules are installed. If so, the domain
is taken from the module that is installed the last.

This should not be case, instead domain should be combined.

task - 3597033

closes odoo/enterprise#61363

X-original-commit: 519e28d96823c9dbe10ec221562daff343391304
Related: odoo/odoo#163138
Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>
Signed-off-by: Sofie Gvaladze (sgv) <sgv@odoo.com>