[FIX] sale_timesheet_enterprise: get timesheets even if no data for leaderboard
Before this commit, when the user goes to Timesheets app, he could get a
traceback because the JS code does manage the case `get_billable_time_target`
method could return an empty when no employee is linked to the current
user.

This commit fixes the issue by managing that case and hide the
leaderboard as we should expect since the billable target is not found
for that current user.

Steps to reproduce the issue
============================
1. Create a new employee
2. Link that employee to a new user
3. Remove the user to the employee (unset Linked user field)
4. Log in as that new user
5. Go to timesheets app

Current Behavior
================
A traceback is occured because the rpc called returned an empty list
instead of a list containing at least one object to get the billable
target for the current user.

Expected Behavior
=================
The grid view of Timesheets app should be loaded as expected and the
leaderboard should not be displayed since no data is found to get the
billable target for that current user.

opw-3862635
opw-3866805
opw-3864353

closes odoo/enterprise#61199

X-original-commit: 0b877857eb3132ff5a896894c7a347a02bfe2bf9
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>