[FIX] account_followup: remove 'CREATE TEMP TABLE'
The readonly flag on `res_partner`.`search` was not respected because
of `CREATE TEMP TABLE followup_data_cache`. It was used to create a
transactional cache. Instead of putting this cache on the database
side, put it on the Python side. On the Python side, the memory spike
should be the same (we've retrieved all the data at some point), but
the memory won't be free until the end of the transaction/cursor.

If the memory problem still occurs, we can easily reduce the memory
usage by 3 by avoiding  the use of dict and dictfetchall usages.

closes odoo/enterprise#61115

X-original-commit: ce2db7362ed7d0f53f974683b1128ff57a94c246
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Rémy Voet (ryv) <ryv@odoo.com>