[FIX] l10n_mx_xml_polizas: xsd validation fetches from internet
Currently, the xsd validation tries to fetch a part of the xsd from
internet.
There is no need for it, we should just have it (as the others).

It causes issues when running the tests if you can't make the request.
(reproducible if you have no internet when running the test).

Linked to runbot error 30521

closes odoo/enterprise#61116

X-original-commit: b04ff1795e7ad74827145bfb320bd4e90c8c5370
Signed-off-by: Habib Ayob (ayh) <ayh@odoo.com>
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>