[FIX] l10n_es_reports: include receipts and ignore entries in modelo347
Currently, receipts are not accounted for in the model 347 report (issue
1). The same happens for invoices for which a payment has been
registered (issue 2).

Setup
-----
* install `l10n_es_reports`
* switch to a Spanish company

Steps to reproduce issue 1
--------------------------
* with a new partner, create and confirm a customer receipt for more
  than €3005.06 (threshold for appearing in the report)
* open the model 347 tax report

We would expect the new partner to be listed in the report, but they
aren't.

Steps to reproduce issue 2
--------------------------
* with a new partner, create and confirm an invoice for more than
  €3005.06
* if you open the model 347 report, you will see the new partner listed
  there
* go back to that invoice and register a payment for it.
* open the model 347 report again

We expect the new partner to still be listed, but they are not.

Cause
-----
issue 1: receipts are missing in the corresponding domains. issue 2: all
account moves are used to calculate the threshold. This means that, for
example, invoices and payments can cancel each other out in the
calculation.

opw-3816370

closes odoo/enterprise#60432

X-original-commit: 54068e7fef8eed380ac1091683a2d01a20cfd227
Signed-off-by: Brice Bartoletti (bib) <bib@odoo.com>
Signed-off-by: Séna Serge Nshimiyimana (sesn) <sesn@odoo.com>