[IMP] *: replace absolute documentation links by widget
Before this commit, in many views the links to the documentation
was a absolute link to a given version of the documentation.
Because of that, the links had to be changed for each version of
Odoo.

After this commit, most of the link to the documentation use a
widget with a relative link that will be adpted with the current
version odoo running on the server.

closes odoo/enterprise#61016

Task-id: 3605445
Related: odoo/odoo#162178
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>