[FIX] documents: fix share and link buttons visibility
Bug
===
Create in this order
- 1 normal document
- 1 link document
- 1 normal document

Select the first and second documents, the link button is not visible.
Select the second and third documents, the link button is visible and shouldn't be.

Similarly, the share button should be visible if at least one document is not archived,
and not if the first document is not active.

Task-3874111

closes odoo/enterprise#61441

X-original-commit: 39d7686a9bc3fc4ebc42d432d1f077c3fbefe7dc
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>