[IMP] project_account_budget: compute the Remaining % for Budgets in Project Updates panel
In the project updates, we can see the margin for the profitability but not for budgets.
It means we can have a clear view of the costs and revenues ratio in % but not for budgets.
After this commit, we will be able to see what has been spent compared to what was allocated in % (in the "Remaining" column) for budgets.

task-3865199

closes odoo/enterprise#61043

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>