[PERF] account_reports: Speed up aged payable report
When a database has lots of companies, it's likely to have a huge number
of rows in the account_partial_reconcile table. Currently, the entire
account_partial_reconcile table is being iterated in the report
generation, only to be filtered out later on in the query. This
superfluous computation is extremely inefficient and it could be easily
avoided by using correct LATERAL JOIN syntax to make use of the filters
applied later on.

Benchmarks:
| account partial reconcile lines | account move lines | Company size | Before |  After |
| --- | ------ | ------ | ------| ----- |
| 588503 | 2339215 | Large | 15m+ (timeout) | ~14.02s |
| 5761 | 27635 | Medium | 15m+ (timeout) |  ~3s  |
| 11 | 1513 | Small | 7.69 | ~1.6s |

Total number of account partial reconcile:
3279438

Total number of account move lines:
10688953

Related ticket:
opw-3819017

closes odoo/enterprise#61401

X-original-commit: 0e5657091bf37ee2734ec9faef3c5ba039435fbf
Signed-off-by: John Laterre (jol) <jol@odoo.com>