[ADD] crm_sale_subscription: update revenues based on the sent quotation
Purpose
=========
Help salespeople keep their opportunities up-to-date by propagating the revenues
they set on Quotations.

Specification
==============
When a quotation is confirmed, if the expected_revenue of the lead is less than the
nonrecurring_total of quote, the expected_revenue of the lead
should be updated to match the nonrecurring total of quote.
If the recurring_revenue of the lead is smaller than the recurring monthly update it with,
Case 1: when no recurring plan is set on lead update it with the recurring monthly.
Case 2 when recurring plan is set on the lead update it with (recurring monthly * number of months)

For example:
If the lead has an expected_revenue of $40 and recurring_revenues of 10$ and
doesn't contain any recurring_plan,

Quote 1: nonrecurring total $45, recurring monthly $12
Then the expected_revenue would be updated from $40 to $45 and
recurring_revenue from $10 to $12.

Task-3806572

closes odoo/enterprise#61550

Related: odoo/odoo#163537
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>