[IMP] knowledge: replace form view's breadcrumbs with hierarchy
Purpose:
--------
Instead of showing the general breadcrumbs, the form view of knowledge now
displays the current article's hierarchy:

< > Root ▸ ... ▸ Parent ▸ Article
          ┌─────────────────┐
          │   Root's child  │
          │ Parent's parent │
          └─────────────────┘

This makes it easier for users to jump to articles (no need to look for
the parent in the sidebar anymore) and to have more context about the
current article.
If there are more than 2 articles in the hierarchy, an ellipsis is shown
between the root and the parent article. This ellipsis opens a dropdown
that will show the other articles in the hierarchy when clicked.
Clicking on any of the article (in the hierarchy and in the dropdown)
opens the article as long as the user has access to that article (articles
that are not accessible will show an error modal), except for the current
article which still allows the user to edit the article's name/icon.

Two new buttons on the left of the hierarchy allow the user to jump
back to previously opened articles, as well as to the previous action
when applicable (like clicking on the previous breadcrumb item if any).
For example if the user is in CRM, inserts the kanban view in article 1,
then opens article 2, he can click on the "back" arrow to go back to the
article 1, then on the "next" arrow to open article 2 again, or click on
the "back" arrow again to come back to the kanban view in CRM.

Technical:
----------
The topbar has 2 new field dependencies: root_article_id and parent_path.
The former is used to display the root in the hierarchy while the latter
is used to determine if the ellipsis should be shown.

Opening an article from an embedded item view, from an article link or
from an article structure now opens the article in the main form view,
instead of loading an action. This allows the new custom breadcrumbs to
keep track of the history of opened articles even if the user opened them
through these behaviors. When an article could not be opened, the dialog
that was shown when an article could not be opened from the article link
behavior will now be shown for all the other cases, instead of redirecting
the user to the knowledge homepage, which might be confusing.

Task-3790413

Part-of: odoo/enterprise#59393