[IMP] knowledge: move add (...) buttons in options dropdown
Purpose:
--------

The "add icon", "add cover" and "add properties" buttons are moved in the
"more options" panel to save space as the breadcrumbs and hierarchy may
take a lot of space in the topbar.

The "add properties" button will now be shown again when one clicked on
it but closed the properties panel without creating any property.

In order to avoid a flicker in the dropdown when adding an icon or a cover,
the "add icon" and "add cover" buttons in the dropdown are replaced with
their "remove icon" and "remove cover" counterparts when an icon and/or
cover is set on the article.

Task-3790413

Part-of: odoo/enterprise#59393