[FIX] website_sale_renting: fix timezone issue in renting
To reproduce :
==============
- from rental settings, make rental unavailable on Sunday
- with a customer, on timezone UTC+X, create a rental order for a product
  that is available on Monday
-> the checkout button is disabled

Problem:
========
the selected dates are converted to UTC before sending them to server,
which makes the server think that the selected date is the day before
the actual selected date.

Solution:
=========
localize the selected dates to the user's timezone as we need to retrieve
the day of the week in the user's timezone.

opw-3778366

closes odoo/enterprise#61047

Signed-off-by: Abdelouahab Laaroussi (abla) <abla@odoo.com>