[FIX] account_sepa: use emp address
**Current behavior:**
If you create a SEPA report for an employee that has address
information in their `hr.employee` record but not their
`res.partner` record, the creditor address will not use the
`hr.employee` information- the corresponding XML node will not
be included.

**Expected behavior:**
The employee address should be in the report if they have one
listed in their HR information.

**Steps to reproduce:**
1. Install the `test_l10n_be_hr_payroll` module, switch to
     'My Belgian Company'

2. In the Employee application, go to an employee (e.g., Bernice
     Jensen) and make sure their listed bank account is verified
     to be able to receive payments

3. In the Payroll application, create a new `To Pay` record for
     the employee from step 2, register it, then create the
     payment report

4. In the Exported File notebook tab, you can see that there is
     no Creditor Address listed in the XML Sepa Report

**Cause of the issue:**
The report builder constructs the creditor address only from
the `res.partner` record.

**Fix:**
Extend the postal address construction such that we get an
address from a partner's employee record (when applicable). The
information should not be mix-and-matched; we will try to use
employee record's address (private addr) if it is available and
accessible, falling back on a partner record address.

opw-3716705

closes odoo/enterprise#62904

X-original-commit: 4ea60de3030ec35d35666683ef82b83a54fad30e
Related: odoo/odoo#166251
Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>
Signed-off-by: Vincent Ethan <etvi@odoo.com>