[FIX] sale_timesheet_enterprise: prevent editing so line if validated timesheet
Steps to reproduce:
-------------------
- create a product:
    - service
    - based on timesheet
    - create on order: project & task

- create a sale order with this product
- confirm the sale order
- go on the created task
- add timesheets with an employee

- go to timesheet
- validate created timesheets

- go to the project settings
- in invoicing tab, change de sale order item for the employee

Issue:
------
The sale order item linked to the validated timesheets are modified.

This behaviour causes negative side effects.
A validated timesheet should no longer be modified,
but the backend logic does, which is not consistent.

Solution:
---------
Make the condition that filters timesheets to update take account
of whether the timesheet is validated or not.

opw-3791062

closes odoo/enterprise#62928

X-original-commit: 699629f668d7e200e1a25748c4d2c0cadf403116
Related: odoo/odoo#166269
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Thomas Lefebvre (thle) <thle@odoo.com>