[FIX] sale_timesheet_enterprise: prevent editing so line if validated timesheet
Steps to reproduce:
-------------------
- create a product:
    - service
    - based on timesheet
    - create on order: project & task

- create a sale order with this product
- confirm the sale order
- go on the created task
- add timesheets with an employee

- go to timesheet
- validate created timesheets

- go to the project settings
- in invoicing tab, change de sale order item for the employee

Issue:
------
The sale order item linked to the validated timesheets are modified.

This behaviour causes negative side effects.
A validated timesheet should no longer be modified,
but the backend logic does, which is not consistent.

Solution:
---------
Make the condition that filters timesheets to update take account
of whether the timesheet is validated or not.

opw-3791062

closes odoo/enterprise#63233

X-original-commit: a24e04b5f12463e095b4ad868a44b4f8001209af
Related: odoo/odoo#166761
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>