[FIX] helpdesk_sale_timesheet: remove commercial partner
Issue:
------
If we remove the sale order line from a helpdesk ticket,
and modify the partner (modify the `is_company` field),
the sale order line will be added back to the ticket.

Solution:
---------
Assuming that this behaviour is due to a limitation in the ORM,
it is necessary to modify the dependency of the compute method.

This fix performs the same correction on the helpdesk side as the commit:
https://github.com/odoo/odoo/commit/5faf5a0dcf4c06716d6a29686d49409b22b05565

opw-3879079

closes odoo/enterprise#61253

X-original-commit: 5cbd24ad11ba17b55f454cd86630a01464dc7904
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>