[FIX] whatsapp: get quality score field while syncing templates
Problem:

The quality_score field is not received in the response while syncing all the
templates. As a result, the default 'none' quality will be applied to all the
templates. And the same goes while syncing a single template. It will only be
updated when Meta posts a request on the webhook to update the template quality.

Solution:

The quality_score field can only be received if we explicitly ask for the
quality_score field in an API request, But the problem is that only the
quality_score field will be received here. Due to this, the API request to sync
a single template and multiple templates are modified, and all the basic fields
like name, components, language, status, category, and ID are included along
with the quality_score field so we receive all basic fields and quality_score
field.

Task-3908894

closes odoo/enterprise#68305

X-original-commit: 7eb1131b413afc2549382d2524ec83445803b682
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>