[FIX] account_accountant: fix reco widget lock on error
Previously, if an error occured while executing execProtectedBankRecAction, the lock on reco actions would never be released.
E.g. some accounts might need a mandatory analytic distribution, and if the user tries to reconcile a move line without it, there would be an userror displayed and the ui would be rendered useless.

X-original-commit: 76061da736b8a1116d07e111e4f1168ae7ab3567
Part-of: odoo/enterprise#61503