[FIX] l10n_us_check_printing: Align check report summary line columns.
Check Report Summary line columns have the correct content, but are  not aligned correctly, with the partner name, date and check sequence number compressed all near the start of the line.

This commit removes the scss style that was causing this problem and fixing a typo for date alignment from 'text-middle' to 'text-center'. We also remove the `xs` infix from the bootstrap classes in the template, since this infix is no longer valid and the correct equivalent is just col-4.

task id: 3708433

closes odoo/enterprise#62359

X-original-commit: 5607baa966d03ddf569efb9d87547515e6ef45ec
Signed-off-by: Quentin De Paoli <qdp@odoo.com>
Signed-off-by: Seif Gneedy (segn) <segn@odoo.com>