[FIX] pos_preparation_display: create lines better
**Current behavior:**
POS preparation lines are currently created in a manner that is
liable to have incorrect quantity information when the same
product appears multiple times on a new order.

**Expected behavior:**
The preparation screen should display complete order data.

**Steps to reproduce:**
1. Install pos_restaurant with demo data

2. In the restaurant pos session, add some productA to the order
     and modify its price with the keypad

3. Add the same product to the order so that the two order lines
     do not merge (as their prices are different)

4. Confirm the order, go to the Kitchen Display, observe that
     there is only 1 preparation line

**Cause of the issue:**
In `_process_preparation_changes()` when we end up with a line
that should have >1 quantity, this value is not correctly
evaluated.

**Fix:**
Iterate over the order lines to create preparation order lines
with accurate quantities.

opw-3857347

closes odoo/enterprise#61155

Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>