[FIX] hr_appraisal_survey: authorize user correctly
We send appraisal survey to employees work address
since https://github.com/odoo/enterprise/pull/52235
Which is desired behavior.

But issue arises here
https://github.com/odoo/odoo/blob/8c853caed272154e74f4f1547891944c8848ff90/addons/survey/controllers/main.py#L88-L90
To authorize user, we check whether answer_sudo.partner_id != request.env.user.partner_id

if employee.user_id.partner_id is not the same as employee.work_contact_id and
we send survey for employee.work_contact_id, they will be redirected to answer_wrong_user

To correctly authorize this user, we override controller and check for this case.
In case user has an employee with work_contact_id that is the same as answer.partner_id,
then we allow them to the survey.

closes odoo/enterprise#62073

X-original-commit: d3054ce9a7514d0706d958213b2ef7c294399193
Signed-off-by: Sofie Gvaladze (sgv) <sgv@odoo.com>
Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>