[IMP] documents: allow collapsing documents inspector
This commit adds the possibility of collapsing the documents inspector
panel in order to have more space to view and preview documents.

It adds a `toggle button`  to the documents inspector. In the uncollapsed
form of  inspector, this button allows to 'fold' the inspector. While in the
collapsed inspector form, the toggle button allows to 'unfold' the
inspector sidebar.

Apart from the 'unfold' button, the collapsed sidebar contains the
'open chatter' button, which works similarly as it does in the uncollapsed
form of inspector.

Technical:
The boolean `isFolded` is introduced to manage the folded/unfolded state of
the documents inspector. On the basis of this boolean, all the inspector sections
are displayed as follows:

`isFolded` is 'True' i.e. the inspector is collapsed:
- It applies the sidebar width and displays only toggle and chatter buttons.
  `d-flex` class is added to sidebar and margin is added to chatter button to
  ensure the buttons are aligned properly.
- Does not display documents preview and info sections.
- Applies the title 'Unfold' to the toggle button and changes its icon to indicate
  the same.

`isFolded` is 'False' i.e. the inspector is uncollapsed:
- It applies the documents inspector width and displays all inspector buttons.
- It now displays both the documents preview, and info sections.
- Applies the title 'Fold' to the toggle button and changes its icon to indicate
  the same.

This commit aligns and adds a margin to both, toggle and chatter buttons
to ensure that there is an appropriate gap between the buttons in uncollpased form
of the inspector.

Task-3741366

closes odoo/enterprise#57128

Signed-off-by: Stéphane Debauche (std) <std@odoo.com>