[FIX] hr_appraisal: determine whether the feedback is readonly
Steps to reproduce:
-------------------
- create an appraisal for an employee;
- be the manager of the appraisal;

Issue:
------
It is not possible to edit feedback before saving it.

Cause:
------
The `accessible_employee_feedback` and `accessible_manager_feedback` fields
are read only if `can_see_employee_publish` and `can_see_manager_publish`
are equal to `False`.
These two fields depend on the `_compute_buttons_display` method,
which will test whether the employees linked to the user are
in the appraisal managers.

The following condition will never be true if the appraisal
has not yet been saved:
```py
user_employees in appraisal.manager_ids
```
This is because manager_ids will be a `NewId` recordset.

Solution:
---------
It is necessary to compare ids and not records directly
because Model does not support equality between
existed records in database and virtual records.

opw-3869446

closes odoo/enterprise#61371

X-original-commit: 644767797c5d42d33536a3bd278865f521c3060a
Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>
Signed-off-by: Thomas Lefebvre (thle) <thle@odoo.com>