[FIX] web_studio: fix undeterministic test
The test edit "one2many form view (2 level) and check chatter allowed" was sometimes failing (https://runbot.odoo.com/runbot/build/61498073)

This was due to the ordering of some rpc calls in two different services. Those services do not depend on one another, so
it is possible that in some situations, their instanciation and the rpc they pass happen in different order.

After this commit this is not a problem because we simply don't assert the rpc "/mail/action".

closes odoo/enterprise#61222

Signed-off-by: Luca Vitali (luvi) <luvi@odoo.com>