[IMP] l10n_be_reports_post_wizard: Kept wizard for BE taxes only
The Belgian "Post a tax report entry" wizard was modified to pop only when the closing entry corresponds to a Belgian tax.

Prior to this commit, the "Post a tax report entry" wizard would pop when attempting to post a closing entry that corresponds to non-Belgian taxes.
It should not pop up.

task-3829124

closes odoo/enterprise#61631

X-original-commit: 8158bc7a3a42941d08ba46e2b5e7d604bb5f5696
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Hesham Saleh (hsal) <hsal@odoo.com>