[IMP] project_forecast, sale_planning: enforce company consistency with tied shifts
After this commit, it is no longer possible to change the company of an SO or a project
if they have been tied to shifts that are not in the company we want to change the SO/project.
One must change the company of the shifts before changing the company of the SO/project.
Ther rationale is that we want to avoid inconsistent situations where the shifts and the SO/project
have different companies.

task-3857690

closes odoo/enterprise#61238

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>