[IMP] *: add support for on-demand quality checks
*: quality_control, quality_mrp, quality_mrp_workorder

This commit adds the ability to create on-demand quality points that
could be used to arbitrarily add quality checks to any transfer/MO. Users
are able to do so via a specific action attached to the
`stock.picking`/`mrp.production` form. On clicking on this action, a
wizard appears where users should specify the product, quality point
and lot/serial number (if applicable). On confirmation, the proper
quality check(s) will be added to the current transfer/MO only.

Task-3870556

Part-of: odoo/enterprise#61237