[FIX] sign: Restrict "Request Signature" action in models
Before this commit: The action of "Request Signature" was shown in weird records accross Odoo.
The idea was to allow adding it in records like tasks, documents, but it is being shown in records that do not have the chatter available to track the signature requests.

This commit aims to fix the issue by adding a new condition for displaying that action, that is:
check if `message_ids` is present on the view fields. since `message_ids` should be available only in chatter.

closes odoo/enterprise#61250

Task: 3861299
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>