[FIX] account_saft: batch dictfetchall to optimize memory usage
Issue -->

The fetchall call made after the querying the dataset in `_saft_fill_report_general_ledger_values`
runs into a memory error if there is a large number of rows.

Solution -->

Use a generator function that takes in the cursor and a batch size as arguments to use `dictfetchmany`
to return rows in batches to optimize memory usage.

opw-3859206

closes odoo/enterprise#62937

X-original-commit: f7e7fbeef45a8846d0dcbf9109d5e59e2c4ed72a
Signed-off-by: Laurent Smet (las) <las@odoo.com>