[FIX] l10n_es_reports: Too many partners excluded for the model 347
Summary
-----
When you try to display the spanish tax report, model 347, if you
have too much partners with few small transactions (the sum is lower
than 3005.06€), an error is raised.

Steps to reproduce
-----
1. Install the l10n_es module
2. Create 30,000 invoices with an amount of 1€, and each having a
distinct partner linked
3. Open the model 347: Accounting > Reporting > Tax Report, and
3select the report called "Tax Report (Mod 347)(ES)" and a date range
3large enough
4. Observe the very long UserError

Cause
-----
Too much partners must be excluded and the list of this partners are
added to a domain which becomes very long and an error is raised when
we use the function ast.literal_eval on it.

Fix
-----
The list of the excluded partners are passed in the "forced_domain"
options. Like that, the ast.literal_eval function is not applied on
it and the error is not raised.

opw-3770944

closes odoo/enterprise#61608

X-original-commit: 490bd2c0e067d3105e30c8b916d5873908a0805f
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>